first pass done
